NLPMetrics

Documentation for NLPMetrics.

NLPMetrics._f_p_r_lcs

NLPMetrics._get_ngrams

NLPMetrics._get_word_ngrams

NLPMetrics._lcs

NLPMetrics._len_lcs

NLPMetrics._recon_lcs

NLPMetrics._split_into_words

NLPMetrics._union_lcs

NLPMetrics.bleu_score

NLPMetrics.get_ngrams

NLPMetrics.rouge

NLPMetrics.rouge_l_sentence_level

NLPMetrics.rouge_l_summary_level

NLPMetrics.rouge_n

NLPMetrics._f_p_r_lcs — Method

_f_p_r_lcs(llcs, m, n)

Computes the LCS-based F-measure score

Arguments:

llcs: Length of LCS

m: number of words in reference summary

n: number of words in candidate summary

Source: (http://research.microsoft.com/en-us/um/people/cyl/download/papers/rouge-working-note-v1.3.1.pdf)

NLPMetrics._get_ngrams — Method

_get_ngrams(n, text)

Calcualtes n-grams. Returns a set of n-grams.

Arguments:

n: provide which n-grams to calculate

text: An array of tokens

NLPMetrics._get_word_ngrams — Method

_get_word_ngrams(n, sentences)

Calculates word n-grams for multiple sentences.

NLPMetrics._lcs — Method

_lcs(x, y)

Utility function to compute the length of the longest common subsequence (lcs) between two strings. The implementation below uses a DP programming algorithm and runs in O(nm) time where n = len(x) and m = len(y).

NLPMetrics._len_lcs — Method

_len_lcs(x, y)

Returns the length of the Longest Common Subsequence between sequences x and y.

NLPMetrics._recon_lcs — Method

_recons_lcs(x, y)

Returns the Longest Subsequence between x and y.

NLPMetrics._split_into_words — Method

_split_into_words(sentences)

Splits multiple sentences into words and flattens the result

NLPMetrics._union_lcs — Method

_union_lcs(evaluated_sentences, reference_sentence)

Returns LCSu(ri, C) which is the LCS score of the union longest common subsequence between reference sentence ri and candidate summary C.

Arguments:

evaluated_sentences: the sentences that have been picked by the summarizer

reference_sentence: one of the sentences in the reference summaries

For example, if ri= w1 w2 w3 w4 w5, and C contains two sentences: c1 = w1 w2 w6 w7 w8 and c2 = w1 w3 w8 w9 w5, then the longest common subsequence of ri and c1 is “w1 w2” and the longest common subsequence of ri and c2 is “w1 w3 w5”. The union longest common subsequence of ri, c1, and c2 is “w1 w2 w3 w5” and LCSu(ri, C) = 4/5.

NLPMetrics.bleu_score — Method

bleu_score(reference_corpus, translation_corpus; max_order=4, smooth=false)

Computes BLEU score of translated segments against one or more references. Returns the BLEU score, n-gram precisions, brevity penalty, geometric mean of n-gram precisions, translationlength and referencelength

Arguments

reference_corpus: list of lists of references for each translation. Each reference should be tokenized into a list of tokens.

translation_corpus: list of translations to score. Each translation should be tokenized into a list of tokens.

max_order: maximum n-gram order to use when computing BLEU score.

smooth=false: whether or not to apply. Lin et al. 2004 smoothing.

NLPMetrics.get_ngrams — Method

get_ngrams(segment, max_order)

Extracts all n-grams upto a given maximum order from an input segment. Returns the counter containing all n-grams upto max_order in segment with a count of how many times each n-gram occurred.

Arguments

segment: text segment from which n-grams will be extracted.

max_order: maximum length in tokens of the n-grams returned by this methods.

NLPMetrics.rouge — Method

rouge(hypotheses, references)

Calculates average rouge scores for a list of hypotheses and references.

NLPMetrics.rouge_l_sentence_level — Method

rouge_l_sentence_level(evaluated_sentences, reference_sentences)

Computes ROUGE-L (sentence level) of two text collections of sentences.

Calculated according to: Rlcs = LCS(X,Y)/m, Plcs = LCS(X,Y)/n, Flcs = ((1 + beta^2)*Rlcs*Plcs) / (Rlcs + (beta^2) * P_lcs)

where: X = reference summary Y = Candidate summary m = length of reference summary n = length of candidate summary

Argumnets:

evaluated_sentences: the sentences that have been picked by the summarizer

reference_sentences: the sentences from the referene set

Source: (http://research.microsoft.com/en-us/um/people/cyl/download/papers/rouge-working-note-v1.3.1.pdf)

NLPMetrics.rouge_l_summary_level — Method

rouge_l_summary_level(evaluated_sentences, reference_sentences)

Computes ROUGE-L (summary level) of two text collections of sentences.

Calculated according to: Rlcs = SUM(1, u)[LCS<union>(ri,C)]/m Plcs = SUM(1, u)[LCS<union>(ri,C)]/n Flcs = ((1 + beta^2)*Rlcs*Plcs) / (Rlcs + (beta^2) * P_lcs)

where: SUM(i,u) = SUM from i through u u = number of sentences in reference summary C = Candidate summary made up of v sentences m = number of words in reference summary n = number of words in candidate summary

Arguments:

evaluated_sentences: the sentences that have been picked by the summarizer

reference_sentence: the sentences in the reference summaries

Source: (http://research.microsoft.com/en-us/um/people/cyl/download/papers/rouge-working-note-v1.3.1.pdf)

NLPMetrics.rouge_n — Method

rouge_n(evaluated_sentences, reference_sentences; n=2)

Computes ROUGE-N of two text collections of sentences. Returns f1, precision, recall for ROUGE-N.

Arguments:

evaluated_sentences: the sentences that have been picked by the summarizer

reference_sentences: the sentences from the referene set

n: size of ngram. Defaults to 2.

Source: (http://research.microsoft.com/en-us/um/people/cyl/download/ papers/rouge-working-note-v1.3.1.pdf)