Classification

Here are some of the metrics for classifications tasks in NLP.

Confusion Matrix

NLPMetrics.confusion_matrix — Function

confusion_matrix(y_pred, y_true)

Function to create a confusionmatrix for classification problems based on provided `ypredandytrue. Expectsytrueandypred`, to be onehotenocded already.

Precision

NLPMetrics.precision — Function

precision(y_pred, y_true; avg_type="macro", sample_weights=nothing)

Computes the precision of the predictions with respect to the labels.

Arguments

y_pred: predicted values.

y_true: ground truth values on the basis of which predicted values are to be assessed. Expects it to be one-hot encoded already

avg_type="macro": Type of average to be used while calculating precision of multiclass models. Can take values as macro, micro and weighted. Default set to macro.

sample_weights: Class weights to be provided when avg_type is set to weighted. Useful in case of imbalanced classes.

Recall

NLPMetrics.recall — Function

recall(y_pred, y_true; avg_type="macro", sample_weights=nothing)

Computes the recall of the predictions with respect to the labels.

Arguments

y_pred: predicted values.

y_true: ground truth values on the basis of which predicted values are to be assessed. Expects it to be one-hot encoded already.

avg_type="macro": Type of average to be used while calculating precision of multiclass models. Can take values as macro, micro and weighted. Default set to macro.

sample_weights: Class weights to be provided when avg_type is set to weighted. Useful in case of imbalanced classes.

Aliases: sensitivity and detection_rate

F_beta Score

NLPMetrics.f_beta_score — Function

f_beta_score(y_pred, y_true; β=1, avg_type="macro", sample_weights=nothing)

Compute f-beta score. The F_beta score is the weighted harmonic mean of precision and recall, reaching its optimal value at 1 and its worst value at 0.

Arguments

y_pred: predicted values.

y_true: ground truth values on the basis of which predicted values are to be assessed. Expects it to be one-hot encoded already

β=1: the weight of precision in the combined score. If β<1, more weight given to precision, while β>1 favors recall.

avg_type="macro": Type of average to be used while calculating precision of multiclass models. Can take values as macro, micro and weighted. Default set to macro.

sample_weights: Class weights to be provided when avg_type is set to weighted. Useful in case of imbalanced classes.