Summarization Metrics

Metric Functions to evaulate models for the task for Summarization.

Rouge Score

NLPMetrics.rouge_n — Function

rouge_n(evaluated_sentences, reference_sentences; n=2)

Computes ROUGE-N of two text collections of sentences. Returns f1, precision, recall for ROUGE-N.

Arguments:

evaluated_sentences: the sentences that have been picked by the summarizer

reference_sentences: the sentences from the referene set

n: size of ngram. Defaults to 2.

Source: (http://research.microsoft.com/en-us/um/people/cyl/download/ papers/rouge-working-note-v1.3.1.pdf)

NLPMetrics.rouge_l_sentence_level — Function

rouge_l_sentence_level(evaluated_sentences, reference_sentences)

Computes ROUGE-L (sentence level) of two text collections of sentences.

Calculated according to: Rlcs = LCS(X,Y)/m, Plcs = LCS(X,Y)/n, Flcs = ((1 + beta^2)*Rlcs*Plcs) / (Rlcs + (beta^2) * P_lcs)

where: X = reference summary Y = Candidate summary m = length of reference summary n = length of candidate summary

Argumnets:

evaluated_sentences: the sentences that have been picked by the summarizer

reference_sentences: the sentences from the referene set

Source: (http://research.microsoft.com/en-us/um/people/cyl/download/papers/rouge-working-note-v1.3.1.pdf)

NLPMetrics.rouge_l_summary_level — Function

rouge_l_summary_level(evaluated_sentences, reference_sentences)

Computes ROUGE-L (summary level) of two text collections of sentences.

Calculated according to: Rlcs = SUM(1, u)[LCS<union>(ri,C)]/m Plcs = SUM(1, u)[LCS<union>(ri,C)]/n Flcs = ((1 + beta^2)*Rlcs*Plcs) / (Rlcs + (beta^2) * P_lcs)

where: SUM(i,u) = SUM from i through u u = number of sentences in reference summary C = Candidate summary made up of v sentences m = number of words in reference summary n = number of words in candidate summary

Arguments:

evaluated_sentences: the sentences that have been picked by the summarizer

reference_sentence: the sentences in the reference summaries

Source: (http://research.microsoft.com/en-us/um/people/cyl/download/papers/rouge-working-note-v1.3.1.pdf)

NLPMetrics.rouge — Function

rouge(hypotheses, references)

Calculates average rouge scores for a list of hypotheses and references.