Translation Metrics

Metric Functions to evaulate models for the task for Machine Translation.

BLEU Score

NLPMetrics.bleu_score — Function

bleu_score(reference_corpus, translation_corpus; max_order=4, smooth=false)

Computes BLEU score of translated segments against one or more references. Returns the BLEU score, n-gram precisions, brevity penalty, geometric mean of n-gram precisions, translationlength and referencelength

Arguments

reference_corpus: list of lists of references for each translation. Each reference should be tokenized into a list of tokens.

translation_corpus: list of translations to score. Each translation should be tokenized into a list of tokens.

max_order: maximum n-gram order to use when computing BLEU score.

smooth=false: whether or not to apply. Lin et al. 2004 smoothing.